+++ src/main/java/com/alkemy/ong/services/MemberService.java
@@ -2,9 +2,9 @@ package com.alkemy.ong.services;
 
 import java.util.List;
 
-import com.alkemy.ong.dtos.ListMemberDTO;
-import com.alkemy.ong.dtos.MemberResponseDTO;
-import com.alkemy.ong.dtos.MemberRequest;
+import com.alkemy.ong.dtos.responses.ListMemberDTO;
+import com.alkemy.ong.dtos.responses.MemberResponseDTO;
+import com.alkemy.ong.dtos.requests.MemberRequest;
 import com.alkemy.ong.entities.Member;
 import com.alkemy.ong.exceptions.DataAlreadyExistException;
 import com.alkemy.ong.exceptions.NotFoundException;
